Separate the Rx layer another module. #5
