+++ collage-gesture-detector-rx/deploy.gradle
@@ -1,0 +1,33 @@
+apply plugin: 'com.novoda.bintray-release'
+
+ext {
+    library_artifactId = 'rx-collage-gesture-detector'
+    library_groupId = 'com.cardinalblue.gesture.rx'
+    library_version = '2.4.4'
+    library_desc = 'The gesture detector used by PicCollage editor.'
+
+    site_url = 'https://github.com/cardinalblue/collage-gesture-detector-android/'
+    vcs_url = 'https://github.com/cardinalblue/collage-gesture-detector-android.git'
+}
+
+publish {
+    bintrayUser = System.getenv('BINTRAY_CBLUE_USER')
+    bintrayKey = System.getenv('BINTRAY_CBLUE_API_KEY')
+
+    repoName = 'android'
+    uploadName = 'rx-collage-gesture-detector'
+
+    groupId = library_groupId
+    artifactId = library_artifactId
+    publishVersion = library_version
+
+    userOrg = System.getenv('BINTRAY_CBLUE_USER')
+    desc = library_desc
+    website = site_url
+    dryRun = false
+}
+
+// Avoid Kotlin docs error
+tasks.withType(Javadoc) {
+    enabled = false
+}
