+++ collage-gesture-detector-rx/src/main/AndroidManifest.xml
@@ -1,0 +1,1 @@
+<manifest package="com.cardinalblue.gesture.rx"/>
