+++ collage-gesture-detector-rx/src/main/java/com/cardinalblue/gesture/rx/GestureEvent.kt
@@ -23,9 +23,10 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-package com.cardinalblue.gesture
+package com.cardinalblue.gesture.rx
 
 import android.graphics.PointF
+import com.cardinalblue.gesture.*
 import java.util.*
 
 sealed class GestureEvent
