+++ collage-gesture-detector-rx/src/main/res/values/strings.xml
@@ -1,0 +1,3 @@
+<resources>
+    <string name="app_name">rx-collage-gesture-detector</string>
+</resources>
