+++ collage-gesture-detector/src/main/AndroidManifest.xml
@@ -1,4 +1,2 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.cardinalblue.gesture">
-    <application/>
-</manifest>
+<manifest package="com.cardinalblue.gesture"/>
